+++ src/index.tsx
@@ -1,5 +1,4 @@
-import { RNDailyTransport } from "./transport";
+import { RNDailyTransport } from './transport';
 import VoiceClientVideoView from './view/VoiceClientVideoView';
 
-export { VoiceClientVideoView, RNDailyTransport }
-
+export { VoiceClientVideoView, RNDailyTransport };