+++ src/transport.ts
@@ -20,7 +20,7 @@ import {
   RTVIClientOptions,
   RTVIMessage,
   RTVIError,
-} from "@pipecat-ai/client-js";
+} from '@pipecat-ai/client-js';
 import { MediaDeviceInfo } from '@daily-co/react-native-webrtc';
 
 export interface DailyTransportAuthBundle {
@@ -32,7 +32,7 @@ export class RNDailyTransport extends Transport {
   // Not able to use declare fields here
   // opened issue: https://github.com/facebook/create-react-app/issues/8918
   private _daily: DailyCall | undefined;
-  private _botId: string = "";
+  private _botId: string = '';
   private _selectedCam: MediaDeviceInfo | Record<string, never> = {};
   private _selectedMic: MediaDeviceInfo | Record<string, never> = {};
   private _selectedSpeaker: MediaDeviceInfo | Record<string, never> = {};
@@ -62,9 +62,9 @@ export class RNDailyTransport extends Transport {
 
     this.attachEventListeners();
 
-    this.state = "disconnected";
+    this.state = 'disconnected';
 
-    console.debug("[RTVI Transport] Initialized");
+    console.debug('[RTVI Transport] Initialized');
   }
 
   get state(): TransportState {
@@ -80,16 +80,14 @@ export class RNDailyTransport extends Transport {
 
   async getAllCams() {
     const { devices } = await this._daily!.enumerateDevices();
-    return devices.filter((d) => d.kind === "videoinput") as MediaDeviceInfo[];
+    return devices.filter((d) => d.kind === 'videoinput') as MediaDeviceInfo[];
   }
 
   updateCam(camId: string) {
-    this._daily!
-      .setCamera(camId)
-      .then(async () => {
-        let inputDevices = await this._daily!.getInputDevices()
-        this._selectedCam = inputDevices.camera as MediaDeviceInfo;
-      });
+    this._daily!.setCamera(camId).then(async () => {
+      let inputDevices = await this._daily!.getInputDevices();
+      this._selectedCam = inputDevices.camera as MediaDeviceInfo;
+    });
   }
 
   get selectedCam() {
@@ -98,16 +96,14 @@ export class RNDailyTransport extends Transport {
 
   async getAllMics() {
     const { devices } = await this._daily!.enumerateDevices();
-    return devices.filter((d) => d.kind === "audio") as MediaDeviceInfo[];
+    return devices.filter((d) => d.kind === 'audio') as MediaDeviceInfo[];
   }
 
   updateMic(micId: string) {
-    this._daily!
-      .setAudioDevice(micId)
-      .then(async () => {
-        let inputDevices = await this._daily!.getInputDevices()
-        this._selectedMic = inputDevices.mic as MediaDeviceInfo;
-      });
+    this._daily!.setAudioDevice(micId).then(async () => {
+      let inputDevices = await this._daily!.getInputDevices();
+      this._selectedMic = inputDevices.mic as MediaDeviceInfo;
+    });
   }
 
   get selectedMic() {
@@ -116,16 +112,14 @@ export class RNDailyTransport extends Transport {
 
   async getAllSpeakers() {
     const { devices } = await this._daily!.enumerateDevices();
-    return devices.filter((d) => d.kind === "audio");
+    return devices.filter((d) => d.kind === 'audio');
   }
 
   updateSpeaker(speakerId: string) {
-    this._daily
-      ?.setAudioDevice(speakerId)
-      .then(async () => {
-        const devicesInUse = await this._daily!.getInputDevices();
-        this._selectedSpeaker = devicesInUse?.speaker;
-      });
+    this._daily?.setAudioDevice(speakerId).then(async () => {
+      const devicesInUse = await this._daily!.getInputDevices();
+      this._selectedSpeaker = devicesInUse?.speaker;
+    });
   }
 
   get selectedSpeaker() {
@@ -185,19 +179,19 @@ export class RNDailyTransport extends Transport {
 
   async initDevices() {
     if (!this._daily) {
-      throw new RTVIError("Transport instance not initialized");
+      throw new RTVIError('Transport instance not initialized');
     }
 
-    this.state = "initializing";
+    this.state = 'initializing';
     await this._daily.startCamera();
     const { devices } = await this._daily.enumerateDevices();
-    const cams = devices.filter((d) => d.kind === "videoinput");
-    const mics = devices.filter((d) => d.kind === "audio");
+    const cams = devices.filter((d) => d.kind === 'videoinput');
+    const mics = devices.filter((d) => d.kind === 'audio');
 
     this._callbacks.onAvailableCamsUpdated?.(cams);
     this._callbacks.onAvailableMicsUpdated?.(mics);
 
-    let inputDevices = await this._daily.getInputDevices()
+    let inputDevices = await this._daily.getInputDevices();
     this._selectedCam = inputDevices.camera;
     this._callbacks.onCamUpdated?.(this._selectedCam as MediaDeviceInfo);
     this._selectedMic = inputDevices.mic;
@@ -209,7 +203,7 @@ export class RNDailyTransport extends Transport {
     if (!this._daily.isRemoteParticipantsAudioLevelObserverRunning())
       await this._daily.startRemoteParticipantsAudioLevelObserver(100);
 
-    this.state = "initialized";
+    this.state = 'initialized';
   }
 
   async connect(
@@ -217,31 +211,31 @@ export class RNDailyTransport extends Transport {
     abortController: AbortController
   ) {
     if (!this._daily) {
-      throw new RTVIError("Transport instance not initialized");
+      throw new RTVIError('Transport instance not initialized');
     }
 
     if (abortController.signal.aborted) return;
 
-    this.state = "connecting";
+    this.state = 'connecting';
 
     try {
       await this._daily.join({
         url: authBundle.room_url,
-        token: authBundle.token,
+        token: authBundle.token || '',
       });
 
       const room = await this._daily.room();
-      if (room && "id" in room && room.config && room.config.exp) {
+      if (room && 'id' in room && room.config && room.config.exp) {
         this._expiry = room.config.exp;
       }
-    } catch (e) {
-      this.state = "error";
-      throw new TransportStartError();
+    } catch (e: Error | any) {
+      this.state = 'error';
+      throw new TransportStartError(e.message);
     }
 
     if (abortController.signal.aborted) return;
 
-    this.state = "connected";
+    this.state = 'connected';
 
     this._callbacks.onConnected?.();
   }
@@ -249,9 +243,9 @@ export class RNDailyTransport extends Transport {
   async sendReadyMessage(): Promise<void> {
     return new Promise<void>((resolve) => {
       (async () => {
-        this._daily!.on("track-started", (ev) => {
+        this._daily!.on('track-started', (ev) => {
           if (!ev.participant?.local) {
-            this.state = "ready";
+            this.state = 'ready';
             this.sendMessage(RTVIMessage.clientReady());
             resolve();
           }
@@ -262,32 +256,32 @@ export class RNDailyTransport extends Transport {
 
   private attachEventListeners() {
     this._daily!.on(
-      "available-devices-updated",
+      'available-devices-updated',
       this.handleAvailableDevicesUpdated.bind(this)
     );
 
     this._daily!.on(
       // TODO, we need to add DailyEventObjectSelectedDevicesUpdated to types overrides inside react-ntive-daily-js
       // @ts-ignore
-      "selected-devices-updated",
+      'selected-devices-updated',
       this.handleSelectedDevicesUpdated.bind(this)
     );
 
-    this._daily!.on("track-started", this.handleTrackStarted.bind(this));
-    this._daily!.on("track-stopped", this.handleTrackStopped.bind(this));
+    this._daily!.on('track-started', this.handleTrackStarted.bind(this));
+    this._daily!.on('track-stopped', this.handleTrackStopped.bind(this));
     this._daily!.on(
-      "participant-joined",
+      'participant-joined',
       this.handleParticipantJoined.bind(this)
     );
-    this._daily!.on("participant-left", this.handleParticipantLeft.bind(this));
-    this._daily!.on("local-audio-level", this.handleLocalAudioLevel.bind(this));
+    this._daily!.on('participant-left', this.handleParticipantLeft.bind(this));
+    this._daily!.on('local-audio-level', this.handleLocalAudioLevel.bind(this));
     this._daily!.on(
-      "remote-participants-audio-level",
+      'remote-participants-audio-level',
       this.handleRemoteAudioLevel.bind(this)
     );
-    this._daily!.on("app-message", this.handleAppMessage.bind(this));
-    this._daily!.on("left-meeting", this.handleLeftMeeting.bind(this));
-    this._daily!.on("nonfatal-error", this.handleNonFatalError.bind(this));
+    this._daily!.on('app-message', this.handleAppMessage.bind(this));
+    this._daily!.on('left-meeting', this.handleLeftMeeting.bind(this));
+    this._daily!.on('nonfatal-error', this.handleNonFatalError.bind(this));
   }
 
   async disconnect() {
@@ -299,12 +293,12 @@ export class RNDailyTransport extends Transport {
   }
 
   public sendMessage(message: RTVIMessage) {
-    this._daily!.sendAppMessage(message, "*");
+    this._daily!.sendAppMessage(message, '*');
   }
 
   private handleAppMessage(ev: DailyEventObjectAppMessage) {
     // Bubble any messages with rtvi-ai label
-    if (ev.data.label === "rtvi-ai") {
+    if (ev.data.label === 'rtvi-ai') {
       this._onMessage({
         id: ev.data.id,
         type: ev.data.type,
@@ -317,10 +311,10 @@ export class RNDailyTransport extends Transport {
     ev: DailyEventObjectAvailableDevicesUpdated
   ) {
     this._callbacks.onAvailableCamsUpdated?.(
-      ev.availableDevices.filter((d) => d.kind === "videoinput")
+      ev.availableDevices.filter((d) => d.kind === 'videoinput')
     );
     this._callbacks.onAvailableMicsUpdated?.(
-      ev.availableDevices.filter((d) => d.kind === "audio")
+      ev.availableDevices.filter((d) => d.kind === 'audio')
     );
   }
 
@@ -344,7 +338,7 @@ export class RNDailyTransport extends Transport {
   }
 
   private handleTrackStarted(ev: DailyEventObjectTrack) {
-    if (ev.type === "screenAudio" || ev.type === "screenVideo") {
+    if (ev.type === 'screenAudio' || ev.type === 'screenVideo') {
       this._callbacks.onScreenTrackStarted?.(
         ev.track,
         ev.participant
@@ -362,7 +356,7 @@ export class RNDailyTransport extends Transport {
   }
 
   private handleTrackStopped(ev: DailyEventObjectTrack) {
-    if (ev.type === "screenAudio" || ev.type === "screenVideo") {
+    if (ev.type === 'screenAudio' || ev.type === 'screenVideo') {
       this._callbacks.onScreenTrackStopped?.(
         ev.track,
         ev.participant
@@ -398,7 +392,7 @@ export class RNDailyTransport extends Transport {
 
     if (p.local) return;
 
-    this._botId = "";
+    this._botId = '';
 
     this._callbacks.onBotDisconnected?.(p);
   }
@@ -415,8 +409,8 @@ export class RNDailyTransport extends Transport {
     for (const participantId in ev.participantsAudioLevel) {
       if (ev.participantsAudioLevel.hasOwnProperty(participantId)) {
         const audioLevel = ev.participantsAudioLevel[participantId];
-        let participant = participants[participantId]
-        if(audioLevel && participant) {
+        let participant = participants[participantId];
+        if (audioLevel && participant) {
           this._callbacks.onRemoteAudioLevel?.(
             audioLevel,
             dailyParticipantToParticipant(participant)
@@ -427,23 +421,22 @@ export class RNDailyTransport extends Transport {
   }
 
   private handleLeftMeeting() {
-    this.state = "disconnected";
-    this._botId = "";
+    this.state = 'disconnected';
+    this._botId = '';
     this._callbacks.onDisconnected?.();
   }
 
   private handleNonFatalError(ev: DailyEventObjectNonFatalError) {
     switch (ev.type) {
-      case "screen-share-error":
+      case 'screen-share-error':
         this._callbacks.onScreenShareError?.(ev.errorMsg);
         break;
     }
   }
 
   get expiry(): number | undefined {
-    return this._expiry
-  };
-
+    return this._expiry;
+  }
 }
 
 const dailyParticipantToParticipant = (p: DailyParticipant): Participant => ({